document outputs
While outputs may still be set even if they are not declared, the use case for this is for outputs whose names are not statically known. Outputs that may not always be set should always be described anyway, as long as their key is known. This enables functionality like autocompletion and linting based on the output names.